requirements-dev.txt
aiokafka==0.10.0
websockets==12.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-asyncio==0.23.6
black==24.4.2
flake8==7.0.0
isort==5.13.2
